third_party/WebKit/LayoutTests/editing/pasteboard/pasting-object.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => assert_selection(
  '<div contenteditable>Hello| World!</div>',
  'insertHTML <object style="border: 1px solid #aaa;"></object>',
  [
    '<div contenteditable>',
      'Hello<object style="border: 1px solid #aaa;"></object>|\u00A0World!',
    '</div>'
  ].join('')),
  'Paste object');
</script>
